requirements.txt
async-timeout==5.0.1
blinker==1.9.0
click==8.3.1
Deprecated==1.3.1
Flask==3.1.2
Flask-Limiter==4.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==25.0
python-dotenv==1.2.1
redis==7.1.0
typing_extensions==4.15.0
Werkzeug==3.1.5
wrapt==2.0.1
